Fix many warnings

+++ src/hash/randomized/TabulationHash.cpp
@@ -6,7 +6,7 @@ TabulationHash::TabulationHash(int k, uint64_t seed) : Hash(k, seed), p(k * 8) {
     len = (int) pow(2, (double) r);
     T = new uint64_t[len * t];
 
-    for (int j1 = 0; j1 < t; j1++) {
+    for (uint64_t j1 = 0; j1 < t; j1++) {
         for (int j2 = 0; j2 < len; j2++) {
             T[j1 * len + j2] = dis(gen);
         }
@@ -32,7 +32,7 @@ void TabulationHash::update(uint8_t c) {
     kmer_hash = (kmer_hash << 8) | c;
 
     hash = 0;
-    for (int j = 0; j < t; j++) {
-        hash ^= T[j * len + (kmer_hash >> (r * j)) & 0xFF]; // TODO instead of FF should be r ones
+    for (uint64_t j = 0; j < t; j++) {
+        hash ^= T[j * len + ((kmer_hash >> (r * j)) & 0xFF)]; // TODO instead of FF should be r ones
     }
 }
